# openapi.yaml
openapi: 3.1.0
...
info:
  title: Gaming Catalog API
  description: Managing the gaming catalog
  version: 0.1.0
paths:
...
  /platforms:
    get:
      tags:
        - Platform
      description: Returns a (filtered, limited, sorted) list of platforms
      parameters:
        - name: name
          in: query
          description: Filter by name of platform
          schema:
            type: string
        - name: sortProperty
          in: query
          description: Sort by platforms property
          schema:
            type: string
            enum: [ id, name ]
            default: id
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortOrder"
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"
        default:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Platform
      summary: Create a new platform
      operationId: createPlatform
      requestBody:
        description: Platform to add to the system
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Platform"
            examples:
              platform:
                id: 1
        required: true
      responses:
        '200':
          description: Platform added to the system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
        '400':
          $ref: "#/components/responses/BadRequest"
        default:
          $ref: "#/components/responses/InternalServerError"
...
  /publishers:
    get:
      tags:
        - Publisher
      description: Returns a (filtered, limited, sorted) list of publishers
      parameters:
        - name: name
          in: query
          description: Filter by name of publisher
          schema:
            type: string
        - name: sortProperty
          in: query
          description: Sort by publisher property
          schema:
            type: string
            enum: [id, name]
            default: id
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortOrder"
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publisher"
        default:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Publisher
      summary: Create a new publisher
      operationId: createPublisher
      requestBody:
        description: Publisher to add to the system
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Publisher"
            examples:
              publisher:
                id: 1
        required: true
      responses:
        '200':
          description: Publisher added to the system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        '400':
          $ref: "#/components/responses/BadRequest"
        default:
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Limit result set
      schema:
        type: integer
    sortOrder:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        $ref: "#/components/schemas/SortOrder"
  responses:
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: internal
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Platform:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        communities:
          schema:
            type: array
            items:
              type: integer
              format: int64
      required:
        - name
    Publisher:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        communities:
          schema:
            type: array
            items:
              type: integer
              format: int64
      required:
        - name
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    SortOrder:
      type: string
      enum: [ asc, desc ]
      default: asc